backend/requirements.txt
Flask
flask-cors
openai  # For LLM interaction (if using OpenAI)
google-generativeai